edges work with container issues
