+++ client/features/prompt-input-page/PromptInputPage.tsx
@@ -4,6 +4,8 @@ import PromptInputForm from '@/client/components/common/PromptInputForm';
 const PromptInputPage = () => {
   return (
     <>
+      <div>PromptInputPage Placeholder:</div>
+      <div>Radix Form Element Placeholder</div>
       <PromptInputForm />
     </>
   );
